Module amazon_sp_shipping::apis::shipping_api

Enums

CancelShipmentError

struct for typed errors of method cancel_shipment

CreateShipmentError

struct for typed errors of method create_shipment

GetAccountError

struct for typed errors of method get_account

GetRatesError

struct for typed errors of method get_rates

GetShipmentError

struct for typed errors of method get_shipment

GetTrackingInformationError

struct for typed errors of method get_tracking_information

PurchaseLabelsError

struct for typed errors of method purchase_labels

PurchaseShipmentError

struct for typed errors of method purchase_shipment

RetrieveShippingLabelError

struct for typed errors of method retrieve_shipping_label

Functions

cancel_shipment

Cancel a shipment by the given shipmentId. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_shipment

Create a new shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_account

Verify if the current account is valid. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_rates

Get service rates. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_shipment

Return the entire shipment object for the shipmentId. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_tracking_information

Return the tracking information of a shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

purchase_labels

Purchase shipping labels based on a given rate. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

purchase_shipment

Purchase shipping labels. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

retrieve_shipping_label

Retrieve shipping label based on the shipment id and tracking id. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.